requirements.txt
bitstring==3.1.9
cryptography==3.4.8
Pillow==8.3.2
streamlit==0.89.0